feat: create stepCount component #103
